refactor(context): split the ranking math out of retrieval.rs
CI's check-file-size guard failed: retrieval.rs reached 1579 lines, over the
1500-line ratchet (#629). The guard is explicit that a file crossing the limit
must be split rather than baselined -- the baseline covers only files that
predate the guard, and adding an entry is what stops the ratchet ratcheting.

The seam is not arbitrary. Everything moved to `retrieval/ranking.rs` is a
pure function of already-loaded candidate metadata: cosine/cosine_blob,
coverage_score, rrf_fuse, dedup_by_content_hash, mmr_select,
budget_tokens_for_bytes, pack_to_budget, dropped_from. None of it touches the
connection, reads a body, or knows what a ContextStore is. The parent keeps
the I/O and the pipeline that sequences the steps.

retrieval.rs 1579 -> 1282; ranking.rs is 323. No behavior change: the moved
code is byte-identical apart from visibility, and `retrieval/tests.rs` reaches
it unchanged through `use super::*`.

All six CI guards, cargo fmt --check, cargo clippy -D warnings, and the full
workspace suite pass locally.
